+++ Data Files/scripts/TamrielData/lib/container_banish.lua
