+++ Data Files/scripts/TamrielData/lib/player_magic_passwall.lua
